Capture errors on the restart step
Capture errors on the restart step by using a script and chaining with the && operator to make sure that such a failure is captured.

Additionally, make the bot is restarted should a failure happen during the time the bot is stopped by capturing to a FAILED variable then propagating the failure result on exit.